Allow fractional position_u for stacking half-U items
Change position_u from int to float so two 0.5U items can occupy
the same U slot at positions 1.0 and 1.5. Migration 007 changes
the column type. Position input uses step=0.5. Validation requires
position_u to be a multiple of 0.5 and >= 1.

+++ backend/app/models.py
@@ -235,7 +235,7 @@ class AppAuthStatus(BaseModel):
 
 
 class RackItemInput(BaseModel):
-    position_u: int
+    position_u: float
     height_u: float = 1
     device_type: str = "other"
     label: str
@@ -248,7 +248,7 @@ class RackItemInput(BaseModel):
 
 class RackItem(BaseModel):
     id: int
-    position_u: int
+    position_u: float
     height_u: float
     device_type: str
     label: str
